Basic navigation & object transforms done; example now displays a simple particle system

+++ src/renderer/mesh.h
@@ -25,7 +25,7 @@ struct PBRMaterialParameters;
 
 struct Vertex {
     glm::vec3 pos;
-    glm::uvec2 uv;
+    glm::vec2 uv;
 };
 
 struct Meshlet {
